+++ lib/src/models/scan_result.g.dart
@@ -6,13 +6,12 @@ part of 'scan_result.dart';
 // JsonSerializableGenerator
 // **************************************************************************
 
-_$ScanResultImpl _$$ScanResultImplFromJson(Map<String, dynamic> json) =>
-    _$ScanResultImpl(
+ScanResult _$ScanResultFromJson(Map<String, dynamic> json) => ScanResult(
       status: $enumDecode(_$ScannerStatusTypeEnumMap, json['status']),
       data: json['scanData'] as String,
     );
 
-Map<String, dynamic> _$$ScanResultImplToJson(_$ScanResultImpl instance) =>
+Map<String, dynamic> _$ScanResultToJson(ScanResult instance) =>
     <String, dynamic>{
       'status': _$ScannerStatusTypeEnumMap[instance.status]!,
       'scanData': instance.data,